Consolidated types into one global import under 'types' folder
