+++ app/src/types/DefinitionsType.ts
@@ -1,6 +1,6 @@
-import { OperationDefinitionType } from './OperationDefinition';
-import { PresetDefinitionType } from './PresetDefinition';
-import { TagDefinitionType } from './TagDefinition';
+import { OperationDefinitionType } from './OperationDefinitionType';
+import { PresetDefinitionType } from './PresetDefinitionType';
+import { TagDefinitionType } from './TagDefinitionType';
 
 export type DefinitionsType = {
   operations: OperationDefinitionType[];
